BJS2-83562 refactoring - kafka and redis listeners separation
