+++ src/main/java/faang/school/analytics/listener/redis/CommentEventListener.java
@@ -1,10 +1,10 @@
-package faang.school.analytics.listener;
+package faang.school.analytics.listener.redis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.analytics.config.redis.RedisProperties;
-import faang.school.analytics.dto.RecommendationReceivedEvent;
-import faang.school.analytics.mapper.event.UserServiceEventMapper;
+import faang.school.analytics.mapper.CommentEventMapper;
 import faang.school.analytics.model.AnalyticsEvent;
+import faang.school.analytics.model.CommentEvent;
 import faang.school.analytics.service.AnalyticsEventService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,26 +16,25 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Set;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
-@Slf4j
-public class RecommendationReceivedEventListener extends AbstractEventListener {
-    private final List<String> topicNameKeys = List.of("recommendation-event");
+public class CommentEventListener extends AbstractEventListener {
+    private final List<String> topicNameKeys = List.of("comment-topic");
     private final RedisProperties properties;
     private final ObjectMapper objectMapper;
-    private final UserServiceEventMapper eventMapper;
-    private final AnalyticsEventService service;
+    private final AnalyticsEventService analyticsEventService;
+    private final CommentEventMapper commentEventMapper;
 
     @Override
     public void onMessage(Message message, byte[] pattern) {
         try {
-            RecommendationReceivedEvent eventDto = objectMapper.readValue(message.getBody(),
-                    RecommendationReceivedEvent.class);
-            AnalyticsEvent event = eventMapper.recommendationReceivedToAnalytics(eventDto);
-            service.saveEvent(event);
-            log.info("Recommendation received event was saved, eventId: {}", event.getId());
+            CommentEvent event = objectMapper.readValue(message.getBody(), CommentEvent.class);
+            AnalyticsEvent analyticsEvent = commentEventMapper.toAnalyticsEvent(event);
+            analyticsEventService.saveEvent(analyticsEvent);
+            log.info("Comment received event was saved, eventId: {}", analyticsEvent.getId());
         } catch (IOException e) {
-            log.error(e.getMessage(), e);
+            log.error("Something went wrong while converting event message to CommentEvent", e);
         }
     }
