+++ src/main/java/faang/school/analytics/listener/redis/PremiumBoughtEventListener.java
@@ -1,10 +1,9 @@
-package faang.school.analytics.listener;
+package faang.school.analytics.listener.redis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.analytics.config.redis.RedisProperties;
-import faang.school.analytics.mapper.CommentEventMapper;
-import faang.school.analytics.model.AnalyticsEvent;
-import faang.school.analytics.model.CommentEvent;
+import faang.school.analytics.dto.premium.PremiumBoughtEvent;
+import faang.school.analytics.mapper.event.UserServiceEventMapper;
 import faang.school.analytics.service.AnalyticsEventService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,22 +18,22 @@ import java.util.Set;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class CommentEventListener extends AbstractEventListener {
-    private final List<String> topicNameKeys = List.of("comment-topic");
+public class PremiumBoughtEventListener extends AbstractEventListener {
+    private final List<String> topicNameKeys = List.of("premium-bought");
     private final RedisProperties properties;
     private final ObjectMapper objectMapper;
-    private final AnalyticsEventService analyticsEventService;
-    private final CommentEventMapper commentEventMapper;
+    private final UserServiceEventMapper userServiceEventMapper;
+    private final AnalyticsEventService service;
 
     @Override
     public void onMessage(Message message, byte[] pattern) {
         try {
-            CommentEvent event = objectMapper.readValue(message.getBody(), CommentEvent.class);
-            AnalyticsEvent analyticsEvent = commentEventMapper.toAnalyticsEvent(event);
-            analyticsEventService.saveEvent(analyticsEvent);
-            log.info("Comment received event was saved, eventId: {}", analyticsEvent.getId());
+            log.debug("Received premium bought event");
+            PremiumBoughtEvent event = objectMapper.readValue(message.getBody(), PremiumBoughtEvent.class);
+            service.saveEvent(userServiceEventMapper.premiumBoughtToAnalytics(event));
+            log.info("Premium bought event for user with id {} analytics successfully saved", event.userId());
         } catch (IOException e) {
-            log.error("Something went wrong while converting event message to CommentEvent", e);
+            log.error(e.getMessage(), e);
         }
     }
